+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayToGameObject.cs
@@ -16,7 +16,7 @@ namespace MimyLab.FukuroUdon
         Inactive,
     }
 
-[HelpURL("https://github.com/mimyquality/FukuroUdon/wiki/Active-Relay#activerelay-to-gameobject")]
+    [HelpURL("https://github.com/mimyquality/FukuroUdon/wiki/Active-Relay#activerelay-to-gameobject")]
     [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay to GameObject")]
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
@@ -28,8 +28,15 @@ namespace MimyLab.FukuroUdon
         private GameObject[] _gameObjects = new GameObject[0];
         [SerializeField]
         private bool _invert = false;
+
+        [Space]
         [SerializeField, Min(0.0f), Tooltip("sec, No delay if 0")]
         private float _delayTime = 0.0f;
+        [SerializeField]
+        private bool _delayLatestOnly = false;
+
+        private int _activateDelayedCount = 0;
+        private int _deactivateDelayedCount = 0;
 
         private void OnEnable()
         {
@@ -40,10 +47,12 @@ namespace MimyLab.FukuroUdon
                 {
                     if (_invert)
                     {
+                        _deactivateDelayedCount++;
                         SendCustomEventDelayedSeconds(nameof(_DeactivateDelayed), _delayTime);
                     }
                     else
                     {
+                        _activateDelayedCount++;
                         SendCustomEventDelayedSeconds(nameof(_ActivateDelayerd), _delayTime);
                     }
                     return;
@@ -62,10 +71,12 @@ namespace MimyLab.FukuroUdon
                 {
                     if (_invert)
                     {
+                        _activateDelayedCount++;
                         SendCustomEventDelayedSeconds(nameof(_ActivateDelayerd), _delayTime);
                     }
                     else
                     {
+                        _deactivateDelayedCount++;
                         SendCustomEventDelayedSeconds(nameof(_DeactivateDelayed), _delayTime);
                     }
                     return;
@@ -77,11 +88,17 @@ namespace MimyLab.FukuroUdon
 
         public void _ActivateDelayerd()
         {
+            _activateDelayedCount--;
+            if (_delayLatestOnly && _activateDelayedCount > 0) { return; }
+
             ToggleActive(true);
         }
 
         public void _DeactivateDelayed()
         {
+            _deactivateDelayedCount--;
+            if (_delayLatestOnly && _deactivateDelayedCount > 0) { return; }
+
             ToggleActive(false);
         }