+++ Packages/com.mimylab.fukuroudon/Runtime/InputFlyingSystem/Scripts/InputFlyingSystem.cs
@@ -29,7 +29,7 @@ namespace MimyLab.FukuroUdon
             {
                 _enableFlight = value;
 
-                if (!value && _isFly)
+                if (!value && _isFlying)
                 {
                     Fly(false);
                 }
@@ -57,10 +57,12 @@ namespace MimyLab.FukuroUdon
 
         // 計算用
         private VRCPlayerApi _localPlayer = null;
-        private float _defaultGravity, _elapsedTime;
-        private bool _isFly, _dampInput;
         private Vector3 _inputDirection = Vector3.zero;
-        private Vector3 _direction, _velocity, _dampVelocity;
+        private bool _isFlying = false;
+        private bool _ignoreGravity = false;
+        private float _defaultGravity;
+        private Vector3 _velocity;
+        private float _elapsedTime = 0.0f;
 
         private void Start()
         {
@@ -69,18 +71,15 @@ namespace MimyLab.FukuroUdon
 
         private void OnDisable()
         {
-            if (_isFly) { Fly(false); }
+            if (_isFlying) { Fly(false); }
         }
 
         private void Update()
         {
-            if (!Utilities.IsValid(_localPlayer)) { return; }
-
-            // 非VRのみキーボード入力受付
-            if (!_localPlayer.IsUserInVR()) { InputKeyboard(); }
+            InputKeyboard();
 
             // 着地したら移動制御を返す
-            if (_isFly && _localPlayer.IsPlayerGrounded())
+            if (_isFlying && _localPlayer.IsPlayerGrounded())
             {
                 Fly(false);
             }
@@ -88,10 +87,8 @@ namespace MimyLab.FukuroUdon
 
         private void FixedUpdate()
         {
-            if (!Utilities.IsValid(_localPlayer)) { return; }
-
             // 飛行処理
-            if (_enableFlight && _isFly)
+            if (_enableFlight && _isFlying)
             {
                 Flight();
             }
@@ -102,6 +99,9 @@ namespace MimyLab.FukuroUdon
         ******************************/
         private void InputKeyboard()
         {
+            // 非VRのみキーボード入力受付
+            if (_localPlayer.IsUserInVR()) { return; }
+
             if (Input.GetKeyDown(fallKeyCode))
             {
                 // 下降
@@ -112,11 +112,31 @@ namespace MimyLab.FukuroUdon
                 // 上昇
                 _inputDirection.y = flipInput ? -1.0f : 1.0f;
             }
-            if (Input.GetKeyUp(fallKeyCode) && !Input.GetKey(riseKeyCode)
-             || Input.GetKeyUp(riseKeyCode) && !Input.GetKey(fallKeyCode))
+            if (Input.GetKeyUp(fallKeyCode))
             {
-                // 上昇中でも下降中でも無くなったら停止
-                _inputDirection.y = 0.0f;
+                if (Input.GetKey(riseKeyCode))
+                {
+                    // 上昇に戻す
+                    _inputDirection.y = flipInput ? -1.0f : 1.0f;
+                }
+                else
+                {
+                    // 上昇中でも下降中でも無くなったら停止
+                    _inputDirection.y = 0.0f;
+                }
+            }
+            if (Input.GetKeyUp(riseKeyCode))
+            {
+                if (Input.GetKey(fallKeyCode))
+                {
+                    // 下降に戻す
+                    _inputDirection.y = flipInput ? 1.0f : -1.0f;
+                }
+                else
+                {
+                    // 上昇中でも下降中でも無くなったら停止
+                    _inputDirection.y = 0.0f;
+                }
             }
         }
 
@@ -140,16 +160,10 @@ namespace MimyLab.FukuroUdon
 
         public override void InputJump(bool value, UdonInputEventArgs args)
         {
-            // 飛行モードへの移行判定
-            if (!EnableFlight) { return; }
-            if (!Utilities.IsValid(_localPlayer)) { return; }
+            if (!_enableFlight) { return; }
 
-            if (_isFly)
-            {
-                // 飛行モード中ならジャンプ中上昇
-                _inputDirection.y = value ? 1.0f : 0.0f;
-            }
-            else if (value && !_localPlayer.IsPlayerGrounded())
+            // 飛行モードへの移行判定
+            if (!_isFlying && value && !_localPlayer.IsPlayerGrounded())
             {
                 // 空中でジャンプしたら飛行モードに移行、移動制御を奪う
                 Fly(true);
@@ -161,7 +175,7 @@ namespace MimyLab.FukuroUdon
         ******************************/
         private void Fly(bool mode)
         {
-            _isFly = mode;
+            _isFlying = mode;
             _localPlayer.Immobilize(mode);
             if (mode)
             {
@@ -169,9 +183,9 @@ namespace MimyLab.FukuroUdon
                 _localPlayer.SetGravityStrength(flightGravity);
 
                 // 飛行計算用変数の初期化
-                _dampInput = false;
-                _velocity = Vector3.zero;
                 _elapsedTime = 0.0f;
+                _ignoreGravity = true;
+                _velocity = _localPlayer.GetVelocity();
             }
             else
             {
@@ -181,39 +195,40 @@ namespace MimyLab.FukuroUdon
 
         private void Flight()
         {
-            // インプットの正規化
-            _direction = Vector3.ClampMagnitude(_inputDirection, 1.0f);
-
-            if (_direction.sqrMagnitude > 0)
+            // インプットを速度に反映
+            var direction = Vector3.ClampMagnitude(_inputDirection, 1.0f);
+            if (direction.sqrMagnitude > 0.0f)
             {
-                if (!_dampInput)
+                // 昇降中は重力を切る
+                if (!_ignoreGravity)
                 {
-                    // 速度減衰フラグ書き戻し
-                    _dampInput = true;
-
-                    // 昇降中は重力を切る
+                    _ignoreGravity = true;
                     _localPlayer.SetGravityStrength(0.0f);
                 }
 
                 _elapsedTime = 0.0f;
-                _velocity = _localPlayer.GetRotation() * _direction * flightSpeed;
+                var rotation = _localPlayer.IsUserInVR() ? _localPlayer.GetRotation() : _localPlayer.GetTrackingData(VRCPlayerApi.TrackingDataType.Head).rotation;
+                _velocity = rotation * direction * flightSpeed;
                 _localPlayer.SetVelocity(_velocity);
+
+                return;
             }
-            else if (_dampInput)
+
+            if (_ignoreGravity)
             {
                 // 最後に入力された速度から徐々に減衰
-                _elapsedTime = Mathf.Clamp01(_elapsedTime + Time.deltaTime);
-                _dampVelocity = Vector3.Lerp(_velocity, Vector3.zero, _elapsedTime / dampTime);
-                _localPlayer.SetVelocity(_dampVelocity);
-
-                if (_elapsedTime >= 1.0f)
+                if (dampTime > 0.0f && _elapsedTime < dampTime)
                 {
-                    // 速度減衰終わり
-                    _dampInput = false;
+                    _elapsedTime += Time.deltaTime;
+                    var velocity = Vector3.Lerp(_velocity, Vector3.zero, _elapsedTime / dampTime);
+                    _localPlayer.SetVelocity(velocity);
 
-                    // 入力が無くなったので飛行中重力に戻す
-                    _localPlayer.SetGravityStrength(flightGravity);
+                    return;
                 }
+
+                // 入力速度が無くなったので飛行中重力に戻す
+                _ignoreGravity = false;
+                _localPlayer.SetGravityStrength(flightGravity);
             }
         }
     }